+++ order/core/src/main/kotlin/com/rarible/protocol/order/core/service/block/nonce/NonceEventSubscriber.kt
@@ -34,7 +34,7 @@ class NonceEventSubscriber(
             changeNonceListener.onNewMakerNonce(
                 history.source.toPlatform(),
                 history.maker,
-                history.newNonce.value.toLong(),
+                history.newNonce.value,
                 history.date
             )
         }
